+++ content/apprenticeships/a100.md
@@ -4,5 +4,8 @@ link: "http://indie-soft.com/a100/"
 location:
   - "New Haven, CT"
 description: "A100 is a software apprenticeship program that provides technical training to aspiring software developers, then matches them with paid internships at local startups."
+roles:
+  - "front-end"
+  - "back-end"
 draft: false
 ---
